feat(indexer): Ensure exporter batches data for both local and s3 implementations (#8)
## Summary

This PR adds config to both local and s3 collector configs to ensure that data sent through is batched just like how the S3ArchiveExporter pipeline builder component does this. It makes sure that if we stream a lot of data through the collector, the files are reasonably sized and also makes sure that we don't send too many files to S3.

## Asana Fixes

https://app.asana.com/1/72322038700732/project/1209091005704648/task/1210882302231695?focus=true

## How to Test

After starting the collector, initiate a stream of data, either from the Tilt function options or otelgen. If using otelgen, a command that will stream data will look like this:

`otelgen --otel-exporter-otlp-endpoint localhost:4317 --insecure --duration 5 --rate 5 logs multi`