fix(docker): Create minio bucket in docker compose, remove create bucket from exporter logic (#12)
## Summary

This PR moves the create bucket requirement out of the OTel collector
exporter and into the startup sequence for docker. Right now, we only
need to create the bucket in local dev contexts. The minio client
[mc](https://github.com/minio/mc) lets us create the bucket once the
minio container is online. By moving it there, we can remove the
exporter code that checks for the bucket and creating it if it does not
exist.

In this manner, we make it a requirement of the exporter to have a
valid, existing bucket for this to work.

## Asana Fixes


https://app.asana.com/1/72322038700732/project/1190453603182737/task/1210891232029777?focus=true

## How to Test

1. Run `tilt up` and confirm that the new `createbucket` service creates
the `my-test-bucket` minio bucket via log message
2. Browse to http://localhost:9001, log into the minio service and
confirm the bucket has been created