feat(indexer): Create protobuf definition for index files and encode them (#14)
## Summary

This builds off the work @maplebed did
[here](https://github.com/honeycombio/enhance-indexing-s3-exporter/pull/13).

This PR creates the protobuf message structure for the contents of the
index file, wires in the exporter to use the marshaling methods
generated, and then adds a utility to unmarshal the protobuf encoded
index files that get uploaded to S3.

## Change summaries

- Creates a `tools/` directory.
- Creates a bash script into `tools/` that uses `protoc` to
auto-generate the protobuf message methods
- Creates a Go utility called `unmarshal-index` which can directly read
protobuf-encoded, gzipped files directly from S3
and report a summary of the contents of the index, including a unique
count of S3 files in the index
- Creates a README that explains how to run `build_proto.sh` as well as
the unmarshal-index utility
- Modifies the `otelcol.Dockerfile` definition and Go workspace which
pulls in more of the project space
  - Adds GOCACHE and GOMODCACHE env vars to the Dockerfile
- Defines the protobuf message structure in `index/protos/index.proto`
- Adds the auto-generated `protoc` code in `index/index.pb.go`
- Modifies the `marshalIndex` and `marshalIndexAsProtobuf` methods to
use the generated protobuf marshal methods.

## Asana Fixes


https://app.asana.com/1/72322038700732/project/1209091005704648/task/1210744461053229?focus=true

## How to Test

1. I would recommend modifying the local exporters config to generate
more files and reducing log output

```
exporters:
  debug:
    #verbosity: detailed
  enhance_indexing_s3_exporter:
    marshaler: otlp_proto
    s3uploader:
      endpoint: 'http://minio:9000/'
      region: 'us-east-1'
      # bucket is created by createbucket service in docker-compose.yml
      s3_bucket: 'my-test-bucket'
      s3_prefix: 'traces-and-logs'
      s3_force_path_style: true
    sending_queue:
        batch:
            flush_timeout: 5s
            max_size: 50000
            min_size: 100
        enabled: true
        queue_size: 500000
        sizer: items
    timeout: 5s
    index:
      enabled: true
      indexed_fields: ['service.instance.id']
```

2. `tilt up`
3. Start the `otelgen_logs_stream` and `otelgen_traces_stream` functions
to stream data into minio
4. Log into the local minio instance at http://localhost:9001/login
5. Wait for index files to land in a minute partition, download an index
file into `tools/`
6. Run the unmarshal-index utility

```
go run unmarshal-index.go index_trace_id_blahblah.binpb.gz
``` 

The unmarshaler should output the contents of the index and report a
summary of the file at the end

```
Summary:
  Field Name: trace_id
  Unique Values: 159
  Number of unique S3 Files: 148
```