refactor: moving a bit of code around to make smaller more testable chunks (#15)
<!-- Lines with HTML comments (like this one) will not be included in
the PR and do not need to be removed -->
<!-- PR titles should follow conventional commits
(https://conventionalcommits.org). For example: `feat(poodle):`,
`feat:`, `fix(retriever):`, `maint(poodle, slos):`, `instr:`. The list
of types we use is at
https://github.com/honeycombio/hound/blob/main/.github/workflows/validate-pr-title.yml
-->

## Summary

Starting in on writing some basic tests for the indexer, this change
pulls logic out of the timer to make it easier to focus a test on that
portion of the code.

As a hint, try looking at the diff with the "ignore whitespace" setting
enabled to more easily see the chunks of code that were just pulled out
into a function (and so indentation changed).

This PR will be one of many addressing the same asana ticket.

<!-- Describe what this PR changes and why. This can be a great place to
include open questions, things you aren’t certain about, tradeoffs made,
other solutions explored & why they weren't appropriate, and specific
areas you want reviewers to focus on. -->

## Asana Fixes


https://app.asana.com/1/72322038700732/project/302048695082307/task/1210932760777316?focus=true

<!-- Paste an Asana task URL to have this PR attached to it -->

## How to Test

No change to tests yet, just getting this out of the way to minimize
merge conflicts.