refactor: add S3Uploader interface so we can swap in a mock for testing (#19)
## Summary

I was looking to add some unit tests to s3writer and pulled this out as
a good way to mock out the actual s3 upload there

## Asana Fixes

<!-- Paste an Asana task URL to have this PR attached to it -->

## How to Test

<!-- Describe the "test plan" for this PR. Guide the reader through a
step-by-step process of how to exercise the code changes, and what they
should expect to see along the way. -->

## Before/After screenshots

<!-- If applicable, provide screenshots of your changes before and
after.--->

### Before

<!-- Provide a screenshot of what the UI looked like before. -->

### After

<!-- Provide a screenshot of what the UI looks like after. -->

## Feature Flag(s) in Use

<!-- LaunchDarkly flags in use (if applicable) -->

## Observability Plan

<!-- What kinds of indicators should we look for to verify that this
code is achieving its goals in production? If appropriate, link a query
result here. -->

## Docs & Rollout Plan

<!-- Link relevant documentation (if available) and describe how this PR
will be rolled out -->