feat(indexer): Prep for v0.0.1 release (#20)
## Summary

This PR prepares the repo so that the indexing exporter's go module can
be tagged with an appropriate semver. It should allow other CI processes
that need to pull in the exporter to do so by its tagged release
version.

It also comments out the detailed verbosities of the OTel collector
configs, allowing the increased verbosity to be something that can be
opted into.

## Asana Fixes


https://app.asana.com/1/72322038700732/project/1209091005704648/task/1211082471768285?focus=true