feat(indexer): Wire in log signal type to use IndexManager to build indexes (#27)
## Summary

This PR follows
https://github.com/honeycombio/enhance-indexing-s3-exporter/pull/24

The prior PR refactored the state management of a shared set of indexes
by created a new `IndexManager` to help manage that state. This PR hooks
in the log signal type to use this `IndexManager` to ensure that logs
get processed by the exporter and added to the same indexes that the
trace signal type is now using.

A new `IndexManager` method is created to facilitate the telemetry
parsing into indexes: `addLogsToIndex`, and now the `consumeLogs` method
of the exporter calls this IndexManager method to have it start
contributing to the index. The unit test covering `consumeLogs` has been
updated and new unit tests have been written to cover `addLogsToIndex`.

## Asana Fixes


https://app.asana.com/1/72322038700732/project/1209091005704648/task/1210744461053240?focus=true

## How to Test

`go test -v ./enhanceindexings3exporter` passes

- Add the index field `k8s.pod.name` to `config/local.yaml` since this
is a field present in the attributes of the otelgen logs generated by
the Tilt function:

```
exporters:
  debug:
    #verbosity: detailed
  enhance_indexing_s3_exporter:
 [...]
    index:
      enabled: true
      indexed_fields: ['service.instance.id', 'k8s.pod.name']
```

- Run `tilt up` at the base of this repo
- Start streams for both logs and traces
- Go to the minio local bucket and download a `k8s.pod.name` index file
and then unmarshal it with the unmarshaling tool (in the `tools/`
directory): `go run unmarshal-index.go
index_k8s.pod.name_aaaa_bbbb_cccc_dddd.binpb.gz`
- Verify that S3 files with `logs_` in the file name are present for
this index.