fix: Change the automatic trace ID index name from trace_id to trace.trace_id (#29)
## Summary

This PR depends on
https://github.com/honeycombio/enhance-indexing-s3-exporter/pull/27
being merged first.

This PR updates how the exporter will automatically index the Trace ID
of telemetry that passes through. What we are learning is that when the
unsampled data is Enhanced from Honeycomb, the trace ID column is
actually called `trace.trace_id` not `trace_id`. There is some kind of
translation happening when data is ingested through Honeycomb that
differs from how OpenTelemetry encodes this in events.

So this PR aligns what the column name should be. The UI work will also
make a corresponding change to ensure that bulk-ingest receives the
automatic trace ID index as the field `trace.trace_id`.

Unit tests have been updated to align with this expectation.

## Asana Fixes


https://app.asana.com/1/72322038700732/project/1209091005704648/task/1211333007221553?focus=true

## How to Test

`go test ./enhanceindexings3exporter`