feat(indexer): disallow file prefix in config (#30)
<!-- Lines with HTML comments (like this one) will not be included in
the PR and do not need to be removed -->
<!-- PR titles should follow conventional commits
(https://conventionalcommits.org). For example: `feat(poodle):`,
`feat:`, `fix(retriever):`, `maint(poodle, slos):`, `instr:`. The list
of types we use is at
https://github.com/honeycombio/hound/blob/main/.github/workflows/validate-pr-title.yml
-->

## Summary
Bulk ingest uses the filename prefix to determine the type of the
telemetry when no filter is provided. If the user specifies the
`file_prefix`, the uploaded telemetry files will have a different
prefix, which will cause error. This PR disables the use of
`file_prefix` in the s3 uploader config.
<!-- Describe what this PR changes and why. This can be a great place to
include open questions, things you aren’t certain about, tradeoffs made,
other solutions explored & why they weren't appropriate, and specific
areas you want reviewers to focus on. -->

## Asana Fixes

https://app.asana.com/1/72322038700732/project/1210733788534383/task/1211429795905462
<!-- Paste an Asana task URL to have this PR attached to it -->

## How to Test
run `go test -v ./enhanceindexings3exporter/`
or add the `file_prefix` in the config and run the collector. It should
still generate files with prefixes of `logs_` or `traces_`. Note that
this is a fallback. In production this should not happen because the
collector is suppose to validate the config.
<!-- Describe the "test plan" for this PR. Guide the reader through a
step-by-step process of how to exercise the code changes, and what they
should expect to see along the way. -->