fix(indexer): Extract automatic indexes from Resource Attribute fields, define precedence (#31)
## Summary

This PR resolves an issue where fields we want to automatically index
like `service.name` are not being indexed correctly. Some of the fields
we want to extract are actually part of a standard list of Attributes
for a Resources (in `service.name`'s case, that Resource is called
[Service](https://opentelemetry.io/docs/specs/semconv/registry/attributes/service/)).

This PR refactors the code to inspect the Attributes of Resources,
Instrumentation Scopes, and Items (Log Records and Spans), and uses a
consistent pattern to correctly detect the attribute values we want to
index. It adheres to the Attribute precedence order set in husky, which
is Item > Scope > Resource.

This should correctly detect the index fields that we were not detecting
properly previously.

## Asana Fixes


https://app.asana.com/1/72322038700732/project/1209091005704648/task/1211348607313280?focus=true

## How to Test

Set `local/config.yaml` to export more trace and log files

```
    sending_queue:
        batch:
            flush_timeout: 5s
            max_size: 1000
            min_size: 50000
        enabled: true
        queue_size: 500000
        sizer: items
    timeout: 5s
```

Run `tilt up`, then start both the logs and traces streams from otelgen.

Observe that the automatic index fields are captured and uploaded to the
local S3 correctly, for `service.name`, `session.id` and
`trace.trace_id`