feat(indexer): Adds management key validation for indexer (#35)
<!-- Lines with HTML comments (like this one) will not be included in
the PR and do not need to be removed -->
<!-- PR titles should follow conventional commits
(https://conventionalcommits.org). For example: `feat(poodle):`,
`feat:`, `fix(retriever):`, `maint(poodle, slos):`, `instr:`. The list
of types we use is at
https://github.com/honeycombio/hound/blob/main/.github/workflows/validate-pr-title.yml
-->

## Summary

<!-- Describe what this PR changes and why. This can be a great place to
include open questions, things you aren’t certain about, tradeoffs made,
other solutions explored & why they weren't appropriate, and specific
areas you want reviewers to focus on. -->
This PR adds validation when a management key is added to the indexer.
In addition to validating the `api_url` and the `api_key` it adds the
team's hcid id that's returned from the response to the indexer config.
It removes making `api_url` and `api_key` optional and now a required
field.

## Asana Fixes

<!-- Paste an Asana task URL to have this PR attached to it -->

https://app.asana.com/1/72322038700732/project/1206994570091048/task/1211499799553955?focus=true

## How to Test

<!-- Describe the "test plan" for this PR. Guide the reader through a
step-by-step process of how to exercise the code changes, and what they
should expect to see along the way. -->
- [x] added unit test
- [x] validating by using a dogfood management key