feat(indexer): Set default batch values in indexer if not present in config (#44)
## Summary

The configuration of the exporter can be set with a `sending_queue`
configuration which governs how to batch telemetry into files before
exporting them to S3:

```
exporters:
  enhance_indexing_s3_exporter:
[...]
    sending_queue:
        batch:
            flush_timeout: 30s
            max_size: 50000
            min_size: 50000
        enabled: true
        queue_size: 500000
        sizer: items
```

But this configuration is optional. If this configuration is not set, we
utilize default constructors in `exporterhelper` to enable the sending
queue and depend on its defaults. Some values are not set by default by
these default constructors. So, this PR modifies the default
configuration slightly to make sure that the sending queue's min and max
size are set to 50,000 items and sets the flush timeout to 30 seconds if
not set. If the configuration is set, then it will use what was set by
the customer.

## Linear Fixes

Resolves COR-2411

## How to Test

Unit tests have been created to test both the default case as well as
the case where the sending queue config is present.