refactor(exporter): update fieldS3Keys to be a map for speedier ops (#43)
## Summary

This PR is a followup to [a
comment](https://github.com/honeycombio/enhance-indexing-s3-exporter/pull/31#discussion_r2382977563)
in a previously merged PR. It should help speed up some operations as we
add/remove s3Keys from indexed field values.

## Linear Fixes


https://linear.app/honeycombio/issue/COR-2317/optimize-fields3keys-to-be-a-map-rather-than-a-slice

## How to Test

Updated tests. There should be no discernible change to how the indexer
works when you run it or the output it has to S3.