docs: Tidy up all README docs in preparation of GA (#53)
## Summary

I did a read-through of the READMEs in this repo and did some tidying up
of the information I saw. The kinds of changes I made were focused on
making sure the information presented is accurate, clarifying bits of
information, and trying to make sure terms and names are used
consistently.

## Linear Fixes

Contributes to COR-2322